Split the padcalculator class
