+++ test/unitTests/viewModelFactories/rowViewModelFactory.test.ts
@@ -1,40 +1,42 @@
 import * as assert from 'assert';
 import { IMock, Mock, It, Times } from "typemoq";
 import { assertExt } from "../../assertExtensions";
-import { PadCalculator } from '../../../src/padCalculator';
+import { PadCalculator } from "../../../src/padCalculation/padCalculator";
 import { RowViewModelFactory } from '../../../src/viewModelFactories/rowViewModelFactory';
 import { Table } from '../../../src/models/table';
 import { Alignment } from '../../../src/models/alignment';
 import { Cell } from '../../../src/models/cell';
 
 suite("RowViewModelFactory.buildRow() tests", () => {
-    let _padCalculator: IMock<PadCalculator>;
+    let _contentPadCalculator: IMock<PadCalculator>;
+    let _separatorPadCalculator: IMock<PadCalculator>;
 
     setup(() => {
-        _padCalculator = Mock.ofType<PadCalculator>();
+        _contentPadCalculator = Mock.ofType<PadCalculator>();
+        _separatorPadCalculator = Mock.ofType<PadCalculator>();
     });
 
     test("padCalculator called for all columns with expected values", () => {
-        const sut = createFactory(_padCalculator.object);
+        const sut = createFactory(_contentPadCalculator.object, _separatorPadCalculator.object);
         const row = 1;
         const table = threeColumnTable();
 
         const rowViewModel = sut.buildRow(row, table);
 
-        _padCalculator.verify(_ => _.getLeftPadding(" ", table, row, 0), Times.once());
-        _padCalculator.verify(_ => _.getLeftPadding(" ", table, row, 1), Times.once());
-        _padCalculator.verify(_ => _.getLeftPadding(" ", table, row, 2), Times.once());
+        _contentPadCalculator.verify(_ => _.getLeftPadding(" ", table, row, 0), Times.once());
+        _contentPadCalculator.verify(_ => _.getLeftPadding(" ", table, row, 1), Times.once());
+        _contentPadCalculator.verify(_ => _.getLeftPadding(" ", table, row, 2), Times.once());
 
-        _padCalculator.verify(_ => _.getRightPadding(" ", table, row, 0), Times.once());
-        _padCalculator.verify(_ => _.getRightPadding(" ", table, row, 1), Times.once());
-        _padCalculator.verify(_ => _.getRightPadding(" ", table, row, 2), Times.once());
+        _contentPadCalculator.verify(_ => _.getRightPadding(" ", table, row, 0), Times.once());
+        _contentPadCalculator.verify(_ => _.getRightPadding(" ", table, row, 1), Times.once());
+        _contentPadCalculator.verify(_ => _.getRightPadding(" ", table, row, 2), Times.once());
     });
 
     test("value returned from padCalculator.getLeftPadding is used to start the row value", () => {
-        const sut = createFactory(_padCalculator.object);
+        const sut = createFactory(_contentPadCalculator.object, _separatorPadCalculator.object);
         const row = 1;
         const table = threeColumnTable();
-        _padCalculator.setup(_ => _.getLeftPadding(It.isAny(), It.isAny(), It.isAny(), It.isAny())).returns(() => "test");
+        _contentPadCalculator.setup(_ => _.getLeftPadding(It.isAny(), It.isAny(), It.isAny(), It.isAny())).returns(() => "test");
         
         const rowViewModel = sut.buildRow(row, table);
 
@@ -42,10 +44,10 @@ suite("RowViewModelFactory.buildRow() tests", () => {
     });
 
     test("value returned from padCalculator.getRightPadding is used to end the row value", () => {
-        const sut = createFactory(_padCalculator.object);
+        const sut = createFactory(_contentPadCalculator.object, _separatorPadCalculator.object);
         const row = 1;
         const table = threeColumnTable();
-        _padCalculator.setup(_ => _.getRightPadding(It.isAny(), It.isAny(), row, It.isAny())).returns(() => "test");
+        _contentPadCalculator.setup(_ => _.getRightPadding(It.isAny(), It.isAny(), row, It.isAny())).returns(() => "test");
         
         const rowViewModel = sut.buildRow(row, table);
 
@@ -53,11 +55,11 @@ suite("RowViewModelFactory.buildRow() tests", () => {
     });
 
     test("Empty middle column is tranformed into a single space", () => {
-        const sut = createFactory(_padCalculator.object);
+        const sut = createFactory(_contentPadCalculator.object, _separatorPadCalculator.object);
         const row = 1;
         const table = threeColumnTableWithEmptyMiddleColumn();
-        _padCalculator.setup(_ => _.getLeftPadding(It.isAny(), It.isAny(), It.isAny(), It.isAny())).returns(() => "L");
-        _padCalculator.setup(_ => _.getRightPadding(It.isAny(), It.isAny(), row, It.isAny())).returns(() => "R");
+        _contentPadCalculator.setup(_ => _.getLeftPadding(It.isAny(), It.isAny(), It.isAny(), It.isAny())).returns(() => "L");
+        _contentPadCalculator.setup(_ => _.getRightPadding(It.isAny(), It.isAny(), row, It.isAny())).returns(() => "R");
         
         const rowViewModel = sut.buildRow(row, table);
 
@@ -67,31 +69,33 @@ suite("RowViewModelFactory.buildRow() tests", () => {
 });
 
 suite("RowViewModelFactory.buildSeparator() tests", () => {
-    let _padCalculator: IMock<PadCalculator>;
+    let _contentPadCalculator: IMock<PadCalculator>;
+    let _separatprPadCalculator: IMock<PadCalculator>;
 
     setup(() => {
-        _padCalculator = Mock.ofType<PadCalculator>();
+        _contentPadCalculator = Mock.ofType<PadCalculator>();
+        _separatprPadCalculator = Mock.ofType<PadCalculator>();
     });
 
     test("padCalculator called for all columns with expected values", () => {
-        const sut = createFactory(_padCalculator.object);
+        const sut = createFactory(_contentPadCalculator.object, _separatprPadCalculator.object);
         const table = threeColumnTable();
 
         const separatorRowViewModel = sut.buildSeparator(table);
 
-        _padCalculator.verify(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), 0), Times.once());
-        _padCalculator.verify(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), 1), Times.once());
-        _padCalculator.verify(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), 2), Times.once());
+        _separatprPadCalculator.verify(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), 0), Times.once());
+        _separatprPadCalculator.verify(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), 1), Times.once());
+        _separatprPadCalculator.verify(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), 2), Times.once());
 
-        _padCalculator.verify(_ => _.getRightPaddingForSeparator("-", It.isAny(), 0), Times.once());
-        _padCalculator.verify(_ => _.getRightPaddingForSeparator("-", It.isAny(), 1), Times.once());
-        _padCalculator.verify(_ => _.getRightPaddingForSeparator("-", It.isAny(), 2), Times.once());
+        _separatprPadCalculator.verify(_ => _.getRightPadding("-", It.isAny(), It.isAny(), 0), Times.once());
+        _separatprPadCalculator.verify(_ => _.getRightPadding("-", It.isAny(), It.isAny(), 1), Times.once());
+        _separatprPadCalculator.verify(_ => _.getRightPadding("-", It.isAny(), It.isAny(), 2), Times.once());
     });
 
     test("value returned from padCalculator.getLeftPadding is used to start the row value", () => {
-        const sut = createFactory(_padCalculator.object);
+        const sut = createFactory(_contentPadCalculator.object, _separatprPadCalculator.object);
         const table = threeColumnTable();
-        _padCalculator.setup(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), 0)).returns(() => "test");
+        _separatprPadCalculator.setup(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), 0)).returns(() => "test");
 
         const separatorRowViewModel = sut.buildSeparator(table);
 
@@ -99,9 +103,9 @@ suite("RowViewModelFactory.buildSeparator() tests", () => {
     });
 
     test("value returned from padCalculator.getRightPadding is used to end the row value", () => {
-        const sut = createFactory(_padCalculator.object);
+        const sut = createFactory(_contentPadCalculator.object, _separatprPadCalculator.object);
         const table = threeColumnTable();
-        _padCalculator.setup(_ => _.getRightPaddingForSeparator(It.isAny(), It.isAny(), It.isAny())).returns(() => "test");
+        _separatprPadCalculator.setup(_ => _.getRightPadding(It.isAny(), It.isAny(), It.isAny(), It.isAny())).returns(() => "test");
 
         const separatorRowViewModel = sut.buildSeparator(table);
 
@@ -109,10 +113,10 @@ suite("RowViewModelFactory.buildSeparator() tests", () => {
     });
 
     test("Empty middle column is tranformed into a single separator dash", () => {
-        const sut = createFactory(_padCalculator.object);
+        const sut = createFactory(_contentPadCalculator.object, _separatprPadCalculator.object);
         const table = threeColumnTableWithEmptyMiddleColumn();
-        _padCalculator.setup(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), It.isAny())).returns(() => "L");
-        _padCalculator.setup(_ => _.getRightPaddingForSeparator("-", It.isAny(), It.isAny())).returns(() => "R");
+        _separatprPadCalculator.setup(_ => _.getLeftPadding("-", It.isAny(), It.isAny(), It.isAny())).returns(() => "L");
+        _separatprPadCalculator.setup(_ => _.getRightPadding("-", It.isAny(), It.isAny(), It.isAny())).returns(() => "R");
 
         const separatorRowViewModel = sut.buildSeparator(table);
 
@@ -141,6 +145,6 @@ function tableFor(rows: string[][]) {
     return table;
 }
 
-function createFactory(padCalculator: PadCalculator): RowViewModelFactory {
-    return new RowViewModelFactory(padCalculator);
+function createFactory(contentPadCalculator: PadCalculator, separatorPadCalculator: PadCalculator): RowViewModelFactory {
+    return new RowViewModelFactory(contentPadCalculator, separatorPadCalculator);
 }
